refactor: api again
