+++ src/api/mod.rs
@@ -1,0 +1,2 @@
+pub mod hero;
+pub mod team;
